chore(refactor): seperate desktop and mobile views
