+++ src/components/common/AdvancedFilter/schema.tsx
@@ -1,0 +1,2 @@
+import { getFilterSchema } from '../../../instance-settings'
+export const schema = getFilterSchema()
