+++ src/components/projects/ProjectsContainer.tsx
@@ -9,7 +9,7 @@ import * as Toast from '../helpers/Toast'
 import { contentType } from '../../types/models'
 import { UserInfo } from '../../modules/account/types'
 import { ProjectsDashboard } from './ProjectsDashboard'
-import ProjectsFilter from '../common/AdvancedFilter/FilterSortButtons'
+import ProjectsFilter from '../common/AdvancedFilter/ProjectsFilter'
 import { deviceWidth } from '../../lib/commonData'
 
 const Container = styled.div`
